+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/networks/cargo/CargoNet.java
@@ -183,6 +183,7 @@ public class CargoNet extends ChestTerminalNetwork {
             }
 
             SlimefunPlugin.getProfiler().newEntry();
+            SlimefunPlugin.getProfiler().scheduleEntries(inputNodes.size());
             Slimefun.runSync(() -> run(inputs, outputs, chestTerminalInputs, chestTerminalOutputs));
         }
     }
@@ -252,12 +253,16 @@ public class CargoNet extends ChestTerminalNetwork {
         // All operations happen here: Everything gets iterated from the Input Nodes.
         // (Apart from ChestTerminal Buses)
         for (Map.Entry<Location, Integer> entry : inputs.entrySet()) {
+            long nodeTimestamp = System.nanoTime();
             Location input = entry.getKey();
             Optional<Block> attachedBlock = getAttachedBlock(input.getBlock());
 
             if (attachedBlock.isPresent()) {
                 routeItems(input, attachedBlock.get(), entry.getValue(), outputs);
             }
+
+            // This will prevent this timings from showing up for the Cargo Manager
+            timestamp += SlimefunPlugin.getProfiler().closeEntry(entry.getKey(), SlimefunItems.CARGO_INPUT_NODE.getItem(), nodeTimestamp);
         }
 
         // Chest Terminal Code