+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/services/profiler/PerformanceSummary.java
@@ -22,6 +22,8 @@ class PerformanceSummary {
 
     // The threshold at which a Block or Chunk is significant enough to appear in /sf timings
     private static final int VISIBILITY_THRESHOLD = 280_000;
+    private static final int MIN_ITEMS = 3;
+    private static final int MAX_ITEMS = 10;
 
     // A minecraft server tick is 50ms and Slimefun ticks are stretched across
     // two ticks (sync and async blocks), so we use 100ms as a reference here
@@ -56,7 +58,7 @@ class PerformanceSummary {
         sender.sendMessage(ChatColor.GOLD + "Performance: " + getPerformanceRating());
         sender.sendMessage("");
 
-        summarizeTimings(totalTickedBlocks + " Blocks", sender, items, entry -> {
+        summarizeTimings(totalTickedBlocks, "block", sender, items, entry -> {
             int count = profiler.getBlocksOfId(entry.getKey());
             String time = NumberUtils.getAsMillis(entry.getValue());
 
@@ -70,24 +72,25 @@ class PerformanceSummary {
             }
         });
 
-        summarizeTimings(chunks.size() + " Chunks", sender, chunks, entry -> {
+        summarizeTimings(chunks.size(), "chunk", sender, chunks, entry -> {
             int count = profiler.getBlocksInChunk(entry.getKey());
             String time = NumberUtils.getAsMillis(entry.getValue());
 
-            return entry.getKey() + " - " + count + "x Blocks (" + time + ")";
+            return entry.getKey() + " - " + count + " block" + (count != 1 ? 's' : "") + " (" + time + ")";
         });
 
-        summarizeTimings(plugins.size() + " Plugins", sender, plugins, entry -> {
+        summarizeTimings(plugins.size(), "plugin", sender, plugins, entry -> {
             int count = profiler.getBlocksFromPlugin(entry.getKey());
             String time = NumberUtils.getAsMillis(entry.getValue());
 
-            return entry.getKey() + " - " + count + "x Blocks (" + time + ")";
+            return entry.getKey() + " - " + count + " block" + (count != 1 ? 's' : "") + " (" + time + ")";
         });
     }
 
-    private void summarizeTimings(String prefix, CommandSender sender, Map<String, Long> map, Function<Map.Entry<String, Long>, String> formatter) {
+    private void summarizeTimings(int count, String name, CommandSender sender, Map<String, Long> map, Function<Map.Entry<String, Long>, String> formatter) {
         Stream<Map.Entry<String, Long>> stream = map.entrySet().stream();
         List<Entry<String, Long>> results = stream.sorted(Map.Entry.comparingByValue(Comparator.reverseOrder())).collect(Collectors.toList());
+        String prefix = count + " " + name + (count != 1 ? 's' : "");
 
         if (sender instanceof Player) {
             TextComponent component = new TextComponent(prefix);
@@ -96,11 +99,14 @@ class PerformanceSummary {
             TextComponent hoverComponent = new TextComponent("  (Hover for details)");
             hoverComponent.setColor(ChatColor.GRAY);
             StringBuilder builder = new StringBuilder();
+
+            int displayed = 0;
             int hidden = 0;
 
             for (Map.Entry<String, Long> entry : results) {
-                if (entry.getValue() > VISIBILITY_THRESHOLD) {
+                if (displayed < MAX_ITEMS && (displayed < MIN_ITEMS || entry.getValue() > VISIBILITY_THRESHOLD)) {
                     builder.append("\n").append(ChatColor.YELLOW).append(formatter.apply(entry));
+                    displayed++;
                 }
                 else {
                     hidden++;
@@ -114,16 +120,19 @@ class PerformanceSummary {
             sender.spigot().sendMessage(component);
         }
         else {
+            int displayed = 0;
             int hidden = 0;
+
             StringBuilder builder = new StringBuilder();
             builder.append(ChatColor.GOLD);
             builder.append(prefix);
             builder.append(ChatColor.YELLOW);
 
             for (Map.Entry<String, Long> entry : results) {
-                if (entry.getValue() > VISIBILITY_THRESHOLD) {
+                if (displayed < MAX_ITEMS && (displayed < MIN_ITEMS || entry.getValue() > VISIBILITY_THRESHOLD)) {
                     builder.append("\n  ");
                     builder.append(ChatColor.stripColor(formatter.apply(entry)));
+                    displayed++;
                 }
                 else {
                     hidden++;