+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/ElectricGoldPan.java
@@ -63,19 +63,14 @@ public abstract class ElectricGoldPan extends AContainer implements RecipeDispla
             if (timeleft > 0 && getSpeed() < 10) {
                 ChestMenuUtils.updateProgressbar(menu, 22, timeleft, processing.get(b).getTicks(), getProgressBar());
 
-                if (ChargableBlock.isChargable(b)) {
-                    if (ChargableBlock.getCharge(b) < getEnergyConsumption()) {
-                        return;
-                    }
+                if (ChargableBlock.getCharge(b) < getEnergyConsumption()) {
+                    return;
+                }
 
-                    ChargableBlock.addCharge(b, -getEnergyConsumption());
-                    progress.put(b, timeleft - 1);
-                }
-                else {
-                    progress.put(b, timeleft - 1);
-                }
+                ChargableBlock.addCharge(b, -getEnergyConsumption());
+                progress.put(b, timeleft - 1);
             }
-            else if (ChargableBlock.isChargable(b)) {
+            else {
                 if (ChargableBlock.getCharge(b) < getEnergyConsumption()) {
                     return;
                 }