+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/abstractItems/AGenerator.java
@@ -17,6 +17,7 @@ import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import io.github.thebusybiscuit.slimefun4.implementation.items.electric.AbstractEnergyProvider;
 import io.github.thebusybiscuit.slimefun4.utils.ChestMenuUtils;
 import io.github.thebusybiscuit.slimefun4.utils.SlimefunUtils;
+import io.github.thebusybiscuit.slimefun4.utils.itemstack.ItemStackWrapper;
 import me.mrCookieSlime.CSCoreLibPlugin.Configuration.Config;
 import me.mrCookieSlime.CSCoreLibPlugin.general.Inventory.ChestMenu.AdvancedMenuClickHandler;
 import me.mrCookieSlime.CSCoreLibPlugin.general.Inventory.ClickAction;
@@ -143,7 +144,7 @@ public abstract class AGenerator extends AbstractEnergyProvider {
             @Override
             public double generateEnergy(Location l, SlimefunItem sf, Config data) {
                 BlockMenu inv = BlockStorage.getInventory(l);
-                boolean chargeable = ChargableBlock.isChargable(l);
+                boolean chargeable = getCapacity() > 0;
                 int charge = chargeable ? ChargableBlock.getCharge(l) : 0;
 
                 if (isProcessing(l)) {
@@ -169,7 +170,7 @@ public abstract class AGenerator extends AbstractEnergyProvider {
                     else {
                         ItemStack fuel = processing.get(l).getInput();
 
-                        if (SlimefunUtils.isItemSimilar(fuel, new ItemStack(Material.LAVA_BUCKET), true) || SlimefunUtils.isItemSimilar(fuel, SlimefunItems.FUEL_BUCKET, true) || SlimefunUtils.isItemSimilar(fuel, SlimefunItems.OIL_BUCKET, true)) {
+                        if (isBucket(fuel)) {
                             inv.pushItem(new ItemStack(Material.BUCKET), getOutputSlots());
                         }
 
@@ -202,6 +203,15 @@ public abstract class AGenerator extends AbstractEnergyProvider {
                 return false;
             }
         };
+    }
+
+    private boolean isBucket(ItemStack item) {
+        if (item == null) {
+            return false;
+        }
+
+        ItemStackWrapper wrapper = new ItemStackWrapper(item);
+        return SlimefunUtils.isItemSimilar(wrapper, new ItemStack(Material.LAVA_BUCKET), true) || SlimefunUtils.isItemSimilar(wrapper, SlimefunItems.FUEL_BUCKET, true) || SlimefunUtils.isItemSimilar(wrapper, SlimefunItems.OIL_BUCKET, true);
     }
 
     private MachineFuel findRecipe(BlockMenu menu, Map<Integer, Integer> found) {